Added VCL cleanup function which return req.body into varnish pipeline
Initial intent was to do it into priv_task.free(), but it is called too late in the request lifecycle. Ideally, we should have a vmod hook which is called "after vcl_recv" for instance. But at the moment, we should either call soap.cleanup() at the end of vcl_recv, or at the beginning of vcl_hash/vcl_pass/vcl_pipe/vcl_synth?.