Rewrite request body processing to use proposed new API
See https://github.com/varnishcache/varnish-cache/pull/3798#issuecomment-1148576467

Changes:

* need std.cache_req_body() now

Unless the request body is only used by the soap vmod (that is, whenever it is
to be sent to the backend also), std.cache_req_body(), optionally with the partial
argument, needs to be used to cache at least as much of the request body
to fulfil any xml lookup used.

Partial caching can be used if, for example, only the XML header is read and the
header is known to be placed at the top.

* VCL is now responsible for checking Content-Encoding

... because it could push filters to support other Content-Encodings

* Varnish Cache supports chunked Encoding for bodies

The test for a body with no Content-Length and no Chunked-Encoding did
not make sense, request bodies could never be closed with "EOL" semantics
(write side shutdown), not even in HTTP/1.0

https://www.rfc-editor.org/rfc/rfc1945.html#section-8.3

   A valid Content-Length is required on all HTTP/1.0 POST requests.

* Polish vcc file

$ABI is implicitly required by varnish-cache as of af87f0cd1a7b8be8a124dca355f04dd641bc5512

The $Module quotes are currently not required.